+++ core/tabs/applications-setup/office-suites/wpsoffice.sh
@@ -3,14 +3,14 @@
 . ../../common-script.sh
 
 installWpsOffice() {
-    if ! command_exists com.wps.Office; then
+    if ! command_exists com.wps.Office && ! command_exists wps; then
         printf "%b\n" "${YELLOW}Installing WPS Office...${RC}"
         case "$PACKAGER" in
             pacman)
                 "$AUR_HELPER" -S --needed --noconfirm wps-office
                 ;;
             *)
-                . ./setup-flatpak.sh
+                checkFlatpak
                 flatpak install flathub com.wps.Office
                 ;;
         esac
